[refactor]판매 등록폼 dto 이동 #68
