+++ src/main/java/wowmarket/wow_server/mypage/myproject/mysales/dto/MySalesDetailResponseDto.java
@@ -1,8 +1,9 @@
-package wowmarket.wow_server.mypage.myproject.dto;
+package wowmarket.wow_server.mypage.myproject.mysales.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import wowmarket.wow_server.domain.Project;
+import wowmarket.wow_server.mypage.myproject.dto.MySalesItemDto;
 
 import java.time.LocalDate;
 import java.util.List;
