+++ src/main/java/wowmarket/wow_server/mypage/myproject/mysales/dto/MySalesListResponseDto.java
@@ -1,4 +1,4 @@
-package wowmarket.wow_server.mypage.myproject.dto;
+package wowmarket.wow_server.mypage.myproject.mysales.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
