+++ src/main/java/wowmarket/wow_server/mypage/myproject/mysales/service/MySalesProjectService.java
@@ -11,10 +11,10 @@ import wowmarket.wow_server.domain.Item;
 import wowmarket.wow_server.domain.Project;
 import wowmarket.wow_server.domain.User;
 import wowmarket.wow_server.global.jwt.SecurityUtil;
-import wowmarket.wow_server.mypage.myproject.dto.MySalesDetailResponseDto;
+import wowmarket.wow_server.mypage.myproject.mysales.dto.MySalesDetailResponseDto;
 import wowmarket.wow_server.mypage.myproject.dto.MySalesFormDto;
 import wowmarket.wow_server.mypage.myproject.dto.MySalesItemDto;
-import wowmarket.wow_server.mypage.myproject.dto.MySalesListResponseDto;
+import wowmarket.wow_server.mypage.myproject.mysales.dto.MySalesListResponseDto;
 import wowmarket.wow_server.repository.*;
 
 import java.util.List;
